backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.20
PyJWT==2.8.0
hashlib-compat==1.0.1